Agrega el componente agregar-al-carrito.component
