+++ src/app/components/agregar-al-carrito/agregar-al-carrito.component.scss
@@ -1,0 +1,53 @@
+@import "../../../colors";
+
+.sic-com-agregar-al-carrito {
+  padding: 0;
+  span.codigo {
+    font-size: 12px;
+    font-weight: normal;
+    color: black;
+  }
+  .descripcion {
+    color: $color-primary;
+    font-weight: bold;
+    margin-top: 0;
+    margin-bottom: 0.5rem;
+  }
+  .input-cantidad {
+    width: 100px;
+    padding: 0 10px;
+  }
+  .en-carrito {
+    margin-top: 0;
+    font-size: 0.85em;
+  }
+  .sic-com-stock {
+    color: white;
+    font-weight: 500;
+  }
+  .sic-com-oferta {
+    display: inline-block;
+    mat-chip {
+      min-height: 22px;
+      font-size: 0.85rem;
+      background: limegreen;
+      color: white;
+    }
+  }
+  .oferta-cantidad {
+    color: limegreen;
+    font-weight: bold;
+    font-size: 0.85em;
+  }
+
+  .precios {
+    color: darkgray;
+  }
+  .seleccionado {
+    font-weight: bold;
+    color: rgba(0, 0, 0, 0.87);
+  }
+  .tachado {
+    text-decoration: line-through;
+  }
+}
